GEODE-4052: Extract GfshExitCodeStatusCommandsTest to multiple acceptance tests for better failure reporting.
